Awesome!  It's working, and it works well.  Definitely took longer than I thought but it's fully functional and can stop/start model imports at any point, and does so in a failsafe way.  Psyched to see how it will all turn out :)
Side note: Python is definitely more versatile than Batch and pretty much better in every way, but I don't know about how I feel between it and Java.  It's cool and all, but it feels so dirty what with the weird "one file contains everything" business.  However, it seems to be the language of the future so I dunno... Also PyCharm isn't that great, it often gives weird suggestions.  Not nearly as good as Rider.